requirements.txt
streamlit
torch
torchvision
numpy
Pillow
faiss-cpu
requests
gitpython
